feat: standalone manifest generation custom columns are editable; empty columns dropped (#183)
* fix: upload timeout error fix

* style: 🎨 fix code style issues with Prettier

* build: windows environment includes certifi

* feat: upload larger datasets without timeout + gen manifests

* style: 🎨 fix code style issues with Prettier

* wip: manifest deletion has no timeout

* style: 🎨 fix code style issues with Prettier

* wip: edit and create new manifests for editing update

* style: 🎨 fix code style issues with Prettier

* chore: remove inaccurate comment

* feat: Penn manifest gen supports existing/new manifest files

* style: manifest editor height capped

If the height isn't capped then a large enough manifest file will push the confirm and cancel buttons to the bottom of the element. This means scrolling is required to exit editing. Now the buttons are always visibile.

* style: 🎨 fix code style issues with Prettier

* WIP: progress bars for manifest UI

* style: 🎨 fix code style issues with Prettier

* wip: fixing an issue

* WIP: refactoring progress containers

* WIP: more modular loading containers

* WIP: progress container in progress

* WIP: modular progress containers

* WIP: demodularizing progress containers

* WIP: first circular progress bar

* feat: import progress bar for manifest files

* refactor: break pennsieve import function into separate file

* feat: progress tracker for generating manifest files

* fix: manifest progress bar disappears on repeat runs

* feat: progress bar resets smoothly

* feat: progress container text updates appropriately

* feat: manifest progress container shows and hides

* feat: continue at step 5 local manifest gen

* wip: preview manifest locally

* feat: local dataset generate manifest outside of original folder

* wip: shell to open manifest files at the generated location

* fix: merge issues

* refactor: formatting timestamps when pulling from Pennsieve

* feat: open folder manifest files were created

* fix: local manifest files will not lose edits on generation

* style: change local manifest gen text

* style: add generate manifest locally to pennsieve man flow

* chore: update explanatory text for Pennsieve manifest gen workflow

* feat: open dialof preview manifest

* feat: preview pennsieve manifest locally in standalone gen

* style: pennsieve loader scrolls into view manifest gen

* style: confirm disappears after ds selection manifest gen

* style: loading spinner text is centered

* style: progress percentage text starts at 0%

* feat: gen man local flow ui functioning

* feat: drop empty manifest columns on gen Pennsieve flow

* style: hide confirm after selecting dataset pennsieve manifest flow

* chore: update incorrect comment

* fix: moving files for preview no longer deletes them

when the files are deleted after previewing it makes it impossible to then upload the files to pennsieve without a change to which path the upload function searches in for the manifest files.

* fix: switch between standard and preview in local manifest gen flow

* style: justify about text in standalone manifest generator

* style: remove success message once manifests have been prepped

* style: generate buttons in separate section manifest generator

* style: hide pennsieve gen section in local workflow

* fix: manifest gen btn section appears after local workflow

* style: smooth scroll to manifest section 5

* chore: manifest generator updated About text to match guided mode

* fix: confirm button pushing into input

* fix: continue button hides and unhides

* style: hide the confirm button on  local ds selection

* style: manifest generator final step turned into a section for pennsieve

* style: section header text child paragraph spaced out

* fix: pennsieve continue btn in local workflow

* WIP: fix spinners in organize

* refactor: organize ds pennsieve spinner uses progress container module

* fix: styling for spinner in local dataset import in Organize

* feat: create manifest folder copies for preview

* chore: add a comment to create duplicate manifest directory function

* chore: remove unused code

* fix: edits on local manifest files brought in to manfiest editor

* chore: update manifest instructional text

* fix: strict equality always false

* feat: only show warning message on generation not preview

* chore: change question text final step of local flow

+++ pyflask/curate/curate.py
@@ -1262,11 +1262,51 @@ def bf_create_new_dataset(datasetname, bf):
         raise e
 
 
-def create_high_level_manifest_files_existing_bf_starting_point(soda_json_structure):
+double_extensions = [
+    ".ome.tiff",
+    ".ome.tif",
+    ".ome.tf2,",
+    ".ome.tf8",
+    ".ome.btf",
+    ".ome.xml",
+    ".brukertiff.gz",
+    ".mefd.gz",
+    ".moberg.gz",
+    ".nii.gz",
+    ".mgh.gz",
+    ".tar.gz",
+    ".bcl.gz",
+]
+
+def get_name_extension(file_name):
+    double_ext = False
+    for ext in double_extensions:
+        if file_name.find(ext) != -1:
+            double_ext = True
+            break
+
+    ext = ""
+    name = ""
+
+    if double_ext == False:
+        name = os.path.splitext(file_name)[0]
+        ext = os.path.splitext(file_name)[1]
+    else:
+        ext = (
+            os.path.splitext(os.path.splitext(file_name)[0])[1]
+            + os.path.splitext(file_name)[1]
+        )
+        name = os.path.splitext(os.path.splitext(file_name)[0])[0]
+    return name, ext
+
+
+def create_high_level_manifest_files_existing_bf_starting_point(soda_json_structure, high_level_folders=["code", "derivative", "docs", "primary", "protocol", "source" ], manifest_progress={}):
     """
     Function to create manifest files for each high-level SPARC folder for an existing Pennsieve dataset.
     Args:
         soda_json_structure: soda dict with information about the dataset to be generated/modified
+        high_level_folders: (optional) list of high-level folders to generate manifests for. Defaults to all primary folders.
+        manifest_progress: (optional) dictionary with information about the progress of the manifest generation. Defaults to empty dictionary.
     Action:
         manifest_files_structure: dict including the local path of the manifest files
     """
@@ -1274,42 +1314,9 @@ def create_high_level_manifest_files_existing_bf_starting_point(soda_json_struct
     manifest_files_structure = {}
     local_timezone = TZLOCAL()
 
-    double_extensions = [
-        ".ome.tiff",
-        ".ome.tif",
-        ".ome.tf2,",
-        ".ome.tf8",
-        ".ome.btf",
-        ".ome.xml",
-        ".brukertiff.gz",
-        ".mefd.gz",
-        ".moberg.gz",
-        ".nii.gz",
-        ".mgh.gz",
-        ".tar.gz",
-        ".bcl.gz",
-    ]
+    global namespace_logger
 
-    def get_name_extension(file_name):
-        double_ext = False
-        for ext in double_extensions:
-            if file_name.find(ext) != -1:
-                double_ext = True
-                break
-
-        ext = ""
-        name = ""
-
-        if double_ext == False:
-            name = os.path.splitext(file_name)[0]
-            ext = os.path.splitext(file_name)[1]
-        else:
-            ext = (
-                os.path.splitext(os.path.splitext(file_name)[0])[1]
-                + os.path.splitext(file_name)[1]
-            )
-            name = os.path.splitext(os.path.splitext(file_name)[0])[0]
-        return name, ext
+    namespace_logger.info("create_high_level_manifest_files_existing_bf_starting_point step 1")
 
     def recursive_folder_traversal(folder, dict_folder_manifest):
         if "files" in folder.keys():
@@ -1333,17 +1340,18 @@ def create_high_level_manifest_files_existing_bf_starting_point(soda_json_struct
                             folder["files"][item]["timestamp"]
                         )
                     elif folder["files"][item]["type"] == "local":
+                        namespace_logger.info("create_high_level_manifest_files_existing_bf_starting_point local files")
                         file_path = folder["files"][item]["path"]
                         filepath = pathlib.Path(file_path)
                         mtime = filepath.stat().st_mtime
                         lastmodtime = datetime.fromtimestamp(mtime).astimezone(
                             local_timezone
                         )
-                        dict_folder_manifest["timestamp"].append(
-                            lastmodtime.isoformat()
-                            .replace(".", ",")
-                            .replace("+00:00", "Z")
-                        )
+
+                        tm = lastmodtime.isoformat().replace(".", ",").replace("+00:00", "Z")
+
+                        namespace_logger.info(tm)
+                        dict_folder_manifest["timestamp"].append(tm)
                 else:
                     dict_folder_manifest["timestamp"].append("")
 
@@ -1371,11 +1379,17 @@ def create_high_level_manifest_files_existing_bf_starting_point(soda_json_struct
 
     dataset_structure = soda_json_structure["dataset-structure"]
 
-    # create local folder to save manifest files temporarly (delete any existing one first)
-    shutil.rmtree(manifest_folder_path) if isdir(manifest_folder_path) else 0
-    makedirs(manifest_folder_path)
+    # create local folder to save manifest files temporarily if the existing files are stale (i.e. not from updating existing manfiest files)
+    if len(high_level_folders) == 6:
+        shutil.rmtree(manifest_folder_path) if isdir(manifest_folder_path) else 0
+        makedirs(manifest_folder_path)
 
     for high_level_folder in list(dataset_structure["folders"]):
+
+        # do not overwrite an existing manifest file 
+        if high_level_folder not in high_level_folders:
+            continue
+
         high_level_folders_present.append(high_level_folder)
 
         folderpath = join(manifest_folder_path, high_level_folder)
@@ -1396,6 +1410,12 @@ def create_high_level_manifest_files_existing_bf_starting_point(soda_json_struct
 
         df = pd.DataFrame.from_dict(dict_folder_manifest)
         df.to_excel(manifestfilepath, index=None, header=True)
+
+        # update the progress of manifest file generation 
+        if manifest_progress != {}:
+            manifest_progress["manifest_files_uploaded"] += 1
+        
+        # add the path to the manifest into the structure
         manifest_files_structure[high_level_folder] = manifestfilepath
 
     return manifest_files_structure
@@ -2015,11 +2035,15 @@ def bf_update_existing_dataset(soda_json_structure, bf, ds):
 
         return
 
-    # Add a new key containing the path to all the files and folders on the
-    # local data structure.
-    # Allows us to see if the folder path of a specfic file already
-    # exists on Pennsieve.
+
     def recursive_item_path_create(folder, path):
+        """
+        Recursively create the path for the item    # Add a new key containing the path to all the files and folders on the
+        local data structure.
+        Allows us to see if the folder path of a specfic file already
+        exists on Pennsieve.
+        """
+        
         if "files" in folder.keys():
             for item in list(folder["files"]):
                 if "folderpath" not in folder["files"][item]:
